Lab 4, halfway

+++ Formes/Forme.cs
@@ -2,7 +2,7 @@
 {
     public interface Forme
     {
-        int Surface();
+        double Surface();
         bool ContientPoint(double x, double y);
 
         BoiteContour BoiteContour();
